+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   * Add resource option & `BakeImages` to all recipes
   * Test that all recipes can receive resources; add match helpers
   * Update main `bake` script for our refactored times
+* Bake `figures in preface` in `college-success`
+
 
 ## [1.32.0] - 2023-06-02